added authorization to patient and appointment routes

+++ models/patient.py
@@ -10,7 +10,7 @@ class Patient(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(25), nullable=False)
     age = db.Column(db.Integer, nullable=False)
-    weight = db.Column(db.Integer, nullable=False)
+    weight = db.Column(db.Float, nullable=False)
     sex = db.Column(db.String(6), nullable=False)
     species = db.Column(db.String, nullable=False)
